FMI MA Hybrid: changes to the MA

+++ ptolemy/actor/lib/fmi/maHybrid/includes/fmi2TypesPlatform.h
@@ -101,7 +101,7 @@
    typedef void*           fmi2FMUstate;                /* Pointer to internal FMU state */
    typedef unsigned int    fmi2ValueReference;
    typedef double          fmi2Real   ;
-   typedef int             fmi2Integer;
+   typedef long            fmi2Integer;
    typedef int             fmi2Boolean;
    typedef char            fmi2Char;
    typedef const fmi2Char* fmi2String;
